+++ confz.d/dynamicDefaults.js
@@ -1,4 +1,9 @@
 module.exports = values => ({
   foo: 'fooFromJs',
-  barf: values.foo + values.bar
+  barf: values.foo + values.bar,
+  a: ['boing', 'bounce'],
+  b: {
+    zzz: 'zzz',
+    yyy: 'yyy'
+  }
 })
