+++ confz.d/schema.yaml
@@ -1,0 +1,9 @@
+description: An awesome schema
+type: object
+properties:
+  foo:
+    type: string
+  bar:
+    type: string
+  a:
+    type: array
