+++ package.json
@@ -11,7 +11,7 @@
     "exec": "docker run bedrocksolutions/confz:latest",
     "pkg": "pkg . --options no-warnings --targets alpine,linux",
     "push": "docker push bedrocksolutions/confz:latest",
-    "test": "node src/index.js --config ./confz.d/confz.yaml --printstack"
+    "test": "node src/index.js run --config ./confz.d/confz.yaml --print-stack"
   },
   "repository": {
     "type": "git",
@@ -25,6 +25,7 @@
   "homepage": "https://github.com/BedrockSolutions/confz#readme",
   "dependencies": {
     "ajv": "^6.10.0",
+    "ajv-merge-patch": "^4.1.0",
     "awaity": "^1.0.0",
     "bluebird": "^3.5.4",
     "chalk": "^2.4.2",
