+++ src/globalConfig.js
@@ -1,6 +1,7 @@
 const { VError } = require('verror')
 
 const { getDirectoryPath, resolvePaths } = require('./fs')
+const { loadFileAtPath } = require('./util')
 const { validate } = require('./validation')
 const { loadFile } = require('./yaml')
 
@@ -28,7 +29,8 @@ const GLOBAL_CONFIG_SCHEMA = {
       },
     },
     valuesSchema: {
-      type: 'object',
+      type: 'string',
+      format: 'uri-reference',
     },
     filterDir: { type: 'string', format: 'uri-reference' },
     resourceDir: { type: 'string', format: 'uri-reference' },
@@ -43,6 +45,7 @@ const ERROR_NAME = 'GlobalConfig'
 const DEFAULT_FILTER_DIR = 'filters'
 const DEFAULT_RESOURCE_DIR = 'resources'
 const DEFAULT_TEMPLATE_DIR = 'templates'
+const DEFAULT_VALUES_SCHEMA = 'schema.yaml'
 
 const getGlobalConfig = async argv => {
   let configFilePath
@@ -55,6 +58,7 @@ const getGlobalConfig = async argv => {
       filterDir: `${homeDir}/${DEFAULT_FILTER_DIR}`,
       resourceDir: `${homeDir}/${DEFAULT_RESOURCE_DIR}`,
       templateDir: `${homeDir}/${DEFAULT_TEMPLATE_DIR}`,
+      valuesSchema: `${homeDir}/${DEFAULT_VALUES_SCHEMA}`,
       valuesExtensions: ['json', 'yml', 'yaml'],
       ...(await loadFile(configFilePath)),
     }
@@ -67,7 +71,7 @@ const getGlobalConfig = async argv => {
         resolvePaths([path])
       ),
       valuesExtensions: prelimGlobalConfig.valuesExtensions,
-      valuesSchema: prelimGlobalConfig.valuesSchema,
+      valuesSchema: await loadFileAtPath(prelimGlobalConfig.valuesSchema),
       defaultValues: await Promise.map(prelimGlobalConfig.defaultValues, path =>
         resolvePaths([path])
       ),
@@ -82,7 +86,7 @@ const getGlobalConfig = async argv => {
         cause,
         name: ERROR_NAME,
         info: {
-          configPath: configFilePath || config,
+          configPath: configFilePath || argv.config,
         },
       },
       `Error initializing global configuration`
